project/requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
Werkzeug
requests
